requirements.txt
python-docx==1.1.2 